Fix: change int to size_t

+++ test/test-parse.cpp
@@ -542,7 +542,7 @@ TEST(test_escape_first_character_in_keyword,
 }
 
 padded_string repeated_code(string8_view before, string8_view inner,
-                            string8_view after, int depth) {
+                            string8_view after, size_t depth) {
   string8 reps;
   reps.reserve((before.size() + after.size()) * depth + inner.size());
   auto append_str_to_reps = [&](string8_view str) {
